refactor(066): move game code into feature layer
