+++ app/src/androidTest/java/org/cescfe/numpairs/ui/screen/GameScreenTestHost.kt
@@ -1,4 +1,4 @@
-package org.cescfe.numpairs.ui.screen
+package org.cescfe.numpairs.feature.game
 
 import androidx.activity.ComponentActivity
 import androidx.compose.runtime.collectAsState
