+++ app/src/main/java/org/cescfe/numpairs/feature/game/GameScreenTokens.kt
@@ -1,4 +1,4 @@
-package org.cescfe.numpairs.ui.screen
+package org.cescfe.numpairs.feature.game
 
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.font.FontWeight
