+++ app/src/main/java/org/cescfe/numpairs/feature/game/GameUiState.kt
@@ -1,4 +1,4 @@
-package org.cescfe.numpairs.ui.screen
+package org.cescfe.numpairs.feature.game
 
 import org.cescfe.numpairs.domain.puzzle.Expression
 import org.cescfe.numpairs.domain.puzzle.OperandSelectionChoice
