+++ app/src/test/java/org/cescfe/numpairs/ui/screen/GameViewModelPresentationStateTest.kt
@@ -1,4 +1,4 @@
-package org.cescfe.numpairs.ui.screen
+package org.cescfe.numpairs.feature.game
 
 import org.cescfe.numpairs.domain.puzzle.OperandSlot
 import org.junit.Assert.assertEquals
